+++ app/src/main/java/com/example/etudiant/CalculMoyenne.java
@@ -8,7 +8,6 @@ import android.widget.TextView;
 import android.content.Intent;
 import android.widget.Toast;
 
-
 import androidx.appcompat.app.AppCompatActivity;
 
 public class CalculMoyenne extends AppCompatActivity {
@@ -19,15 +18,14 @@ public class CalculMoyenne extends AppCompatActivity {
     Button boutonConfirmer;
 
     // Variables to store global calculation
-    double sommeNotes = 0;
-    double sommeCoef = 0;
+    float sommeNotes = 0;
+    float sommeCoef = 0;
     String cin;
-
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_calcul_moyenne); // change if XML name is different
+        setContentView(R.layout.activity_calcul_moyenne);
 
         // Linking views
         nomMatiere = findViewById(R.id.nomMatiere);
@@ -41,7 +39,6 @@ public class CalculMoyenne extends AppCompatActivity {
         boutonAjout = findViewById(R.id.boutonAjout);
         boutonAnnuler = findViewById(R.id.boutonAnnuler);
         boutonConfirmer = findViewById(R.id.boutonConfirmer);
-
 
         // Ajouter Matiere
         boutonAjout.setOnClickListener(new View.OnClickListener() {
@@ -58,10 +55,10 @@ public class CalculMoyenne extends AppCompatActivity {
                     return;
                 }
 
-                double tpVal = Double.parseDouble(tp.getText().toString());
-                double dsVal = Double.parseDouble(ds.getText().toString());
-                double examVal = Double.parseDouble(exam.getText().toString());
-                double coefVal = Double.parseDouble(coef.getText().toString());
+                float tpVal = Float.parseFloat(tp.getText().toString());
+                float dsVal = Float.parseFloat(ds.getText().toString());
+                float examVal = Float.parseFloat(exam.getText().toString());
+                float coefVal = Float.parseFloat(coef.getText().toString());
 
                 if (coefVal < 1 || coefVal > 2) {
                     Toast.makeText(CalculMoyenne.this, "Le coefficient doit être entre 1 et 2", Toast.LENGTH_SHORT).show();
@@ -76,15 +73,14 @@ public class CalculMoyenne extends AppCompatActivity {
                     return;
                 }
 
-
                 // Score of one subject
-                double scoreMatiere = (tpVal * 0.20) + (dsVal * 0.10) + (examVal * 0.70);
+                float scoreMatiere = (tpVal * 0.20f) + (dsVal * 0.10f) + (examVal * 0.70f);
 
                 // Update totals
                 sommeNotes += scoreMatiere * coefVal;
                 sommeCoef += coefVal;
 
-                double moyenne = sommeNotes / sommeCoef;
+                float moyenne = sommeNotes / sommeCoef;
 
                 tvMoyenne.setText("Moyenne : " + String.format("%.2f", moyenne));
 
@@ -100,7 +96,6 @@ public class CalculMoyenne extends AppCompatActivity {
             }
         });
 
-
         boutonConfirmer.setOnClickListener(v -> {
 
             if (sommeCoef == 0) {
@@ -108,7 +103,7 @@ public class CalculMoyenne extends AppCompatActivity {
                 return;
             }
 
-            double moyenneFinale = sommeNotes / sommeCoef;
+            float moyenneFinale = sommeNotes / sommeCoef;
 
             BaseDeDonnees db = new BaseDeDonnees(this);
             Eleve eleve = db.chercherEleveParCin(cin);
@@ -132,9 +127,7 @@ public class CalculMoyenne extends AppCompatActivity {
             }
         });
 
-
-
-        Button btnRetour=findViewById(R.id.boutonRetour);
+        Button btnRetour = findViewById(R.id.boutonRetour);
         btnRetour.setOnClickListener(v -> finish());
     }
